update devapp
move nomap and other templates into devapp directory,
add basic form which uses x-for to apply foreign key
values into the form for submission, update excludes_list
to use values directly from model choice class
