+++ geodata_mart/devapp/templates/devapp/links.html
@@ -35,5 +35,8 @@
   <li>
     <a href="{% url 'devapp:map' %}">Development Map Form</a>
   </li>
+  <li>
+    <a href="{% url 'devapp:nomap' "1" %}">Development NoMap Form for Project 1</a>
+  </li>
 </ul>
 {% endblock content %}
